+++ app/layout.js
@@ -9,6 +9,11 @@ import { store } from "../store/store";
 
 const inter = Inter({ subsets: ["latin"] });
 
+export const metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
